pyproject.toml
[project]
name = "messaging-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.4.3",
    "aioboto3>=15.2.0",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "bcrypt<4",
    "dependency-injector>=4.48.2",
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "openpyxl>=3.1.5",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "python-socks[asyncio]>=2.8.0",
    "qrcode[pil]>=8.2",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.43",
    "telethon>=1.42.0",
    "uvicorn[standard]>=0.37.0",
]
